src/requirements.txt
pygithub==1.46
